requirements.txt
pyalsaaudio==0.8.4
django==1.11.6
djangorestframework==3.5.4
django-rest-swagger==2.1.2
